fixed host.json resolution
