+++ custom-backends/dotnetIsolated-mssql/nuspec.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
     <metadata minClientVersion="2.5">
         <id>DurableFunctionsMonitor.DotNetIsolated.MsSql</id>
-        <version>6.6.0-beta3</version>
+        <version>6.6.0-beta4</version>
         <requireLicenseAcceptance>false</requireLicenseAcceptance>
         <description>DurableFunctionsMonitor.DotNetIsolated.MsSql</description>
         <authors>DurableFunctionsMonitor</authors>
